move themes into a feature and cookie helpers to lib (#210)
* move themes into a feature and cookie helpers to lib

* add a theme constants file

* add a theme constants file

* only export what's necessary from index

* themes: seperate server from client code
