+++ app/features/theme/index.ts
@@ -1,0 +1,2 @@
+export { useTheme } from './use-theme';
+export { ThemeProvider } from './theme-provider';
