+++ app/features/theme/use-theme.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeContext } from './theme-provider';
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+}
